requirements.txt
Flask==0.12.2
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
python-dateutil==2.7.2
requests==2.18.4
six==1.11.0
urllib3==1.22
